feat: добавить repository

+++ stats/dto/src/main/java/ru/practicum/ewm/dto/ViewStatsDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class ViewStatsDto {
     String app;
     String uri;
-    Integer hits;
+    Long hits;
 }
